Refactor neighbors-based metrics to use `NeighborsResults` (#129)
Refactor neighbors-based metrics to use NeighborsResults, making it simpler for users as there is no confusion over passing distances or connectivities. Also reduces overhead of converting back and forth between sparse and dense representations of the neighbor results.
Compute umap-based connectivities in this package instead of scanpy, and make the implementation more efficient
Remove unnecessary vmap in pcr regression
Update notebooks
Drop usage of private scanpy functions. Scanpy can likely be dropped as a dependency in a future release if pca with implicit centering for sparse matrices makes it into sklearn
Bump to 0.5.0
Fixes #109 as lisi always uses the dense (indices, distances) representation for neighbors. The issue arose when the approximate neighbors method gave a distance of zero to non-self cells.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jan Engelmann <jan.engelmann@helmholtz-munich.de>